requirements.txt
python-dotenv
langchain
langchain-groq
langchain_community
langchain_core
streamlit